[E-MEMO-RETIRE S3-2] Backend router memos.py + eventbus memo 이벤트 제거 (#928)
* feat(backend): E-MEMO-RETIRE S3-2 — routers/memos.py + eventbus 제거

- app/routers/memos.py 삭제 (890줄)
- app/services/eventbus.py 삭제 (dispatch_memo_event 전용)
- app/services/event_taxonomy.py: memo_created + memo.reply_created 이벤트 정의 제거
- app/main.py: memos import + include_router 제거
- app/routers/workflow_report.py: _fire_webhook 인라인 (memos 의존성 제거)


* [E-MEMO-RETIRE S3-1] MCP 메모 도구 제거 — chat 도구 보존 (#927)

* feat(mcp): E-MEMO-RETIRE S3-1 — MCP 메모 도구 제거, chat 도구 보존

- tools/memos.py 삭제
- tools/chat.py 신설 (send_chat_message, create_conversation, list_chat_messages 3개 보존)
- server.py: memos import → chat import, 메모 도구 7개 등록 제거


* test(mcp): E-MEMO-RETIRE S3-1 — test_contract 메모 도구 7종 제거 반영

- 도구 수 88 → 86
- EXPECTED_TOOLS에서 memo 7종 제거, chat (3)으로 업데이트
- test_project_id_not_in_schema, test_optional_fields, test_flat_schema에서 list_memos 참조 제거


* test(mcp): E-MEMO-RETIRE S3-1 — test_e2e_dev 메모 도구 제거 반영

- test_tools_list_88_tools → test_tools_list_86_tools (86으로 수정)
- READ_ONLY_TOOLS에서 sprintable_list_memos, sprintable_list_my_memos 제거


---------


* test(backend): E-MEMO-RETIRE S3-2 — test_event_taxonomy memo 테스트 제거

- memos.py 삭제로 인한 ModuleNotFoundError 해소
- _infer_trigger_type, _resolve_author_role import 제거
- memo_created, memo.reply_created 관련 테스트 케이스 제거
- TestInferTriggerType, TestResolveAuthorRole 클래스 제거


* test(backend): E-MEMO-RETIRE S3-2 — 삭제 모듈 참조 테스트 파일 제거

- tests/test_eventbus_s4.py 삭제 (dispatch_memo_event 전용)
- tests/test_memo_02.py 삭제 (_parse_entity_embeds 전용)


---------
